Move tests to NameHelperTest
The Query's in question are located in ShowNameHelper. All of them are used in ShowName (but only some are used in ShowObservation.)
